【问题单修改】修复文件复制
【问题单修改】修复文件复制

+++ pkg/deploy/action/storage_operate/dr_deploy_operate/dr_deploy.py
@@ -1152,9 +1152,13 @@ class DRDeploy(object):
             encrypted_pwd = KmcResolve.kmc_resolve_password("encrypted", self.dm_passwd)
             self.record_disaster_recovery_info("dm_pwd", encrypted_pwd)
             os.chmod(os.path.join(CURRENT_PATH, "../../../config/dr_deploy_param.json"), mode=0o644)
-            shutil.copy(os.path.join(CURRENT_PATH, "../../../config/dr_deploy_param.json"), "/opt/cantian/config/")
+            dst_file = "/opt/cantian/config/dr_deploy_param.json"
+            if not os.path.exists(dst_file):
+                shutil.copy(os.path.join(CURRENT_PATH, "../../../config/dr_deploy_param.json"), "/opt/cantian/config/")
             share_path = f"/mnt/dbdata/remote/metadata_{self.dr_deploy_info.get('storage_metadata_fs')}"
-            shutil.copy(os.path.join(CURRENT_PATH, "../../../config/dr_deploy_param.json"), share_path)
+            dst_share_file = f"{share_path}/dr_deploy_param.json"
+            if not os.path.exists(dst_share_file):
+                shutil.copy(os.path.join(CURRENT_PATH, "../../../config/dr_deploy_param.json"), share_path)
             self.restart_cantian_exporter()
         try:
             _execute()
